Change Offset API field type from uint64 to int64
Kubernetes API conventions prohibit uint64 because JSON numbers are
IEEE 754 doubles with only 53 bits of integer precision. The
structured-merge-diff library deliberately excludes uint64, causing a
panic during server-side apply operations.

Change the Offset field in kprobe and uprobe attach info structs from
uint64 to int64. The generated CRD schema already used format: int64
so no manifest changes are required. Add explicit uint64 casts at the
gRPC boundary where the bpfman protobuf definition expects uint64.

+++ controllers/bpfman-agent/cl_kprobe_program.go
@@ -95,7 +95,7 @@ func (r *ClKprobeProgramReconciler) getAttachRequest() *gobpfman.AttachRequest {
 			Info: &gobpfman.AttachInfo_KprobeAttachInfo{
 				KprobeAttachInfo: &gobpfman.KprobeAttachInfo{
 					FnName:   r.currentLink.Function,
-					Offset:   r.currentLink.Offset,
+					Offset:   uint64(r.currentLink.Offset),
 					Metadata: map[string]string{internal.UuidMetadataKey: string(r.currentLink.UUID)},
 				},
 			},
